Target: Introduce namespace descriptions.
If a file named `Description.md` is present in the namespace directory,
then it is used as the namespace description.

To match a namespace to its directory, PSR-4 mappings are required. Thus
`kitab compile` comes with the new `--with-composer` option, defaulting
to `./composer.json`, to read these mappings from the Composer file of
your project.