Add PHAR support.
And it was totally crazy.

Generating the PHAR is easy. The stub is easy. However, `atoum` is
inside the PHAR. We run `atoum` from a shell (now directly from PHP, but
still from a shell). A shell cannot load a file inside a PHAR, that's
obvious. The idea is to expose `atoum` through a “private” API:
`kitab.phar atoum`.

So now, Kitab in a PHAR can execute atoum. However, we have a big issue
with autoloaders. Kitab needs its own autoloader that is inside the
PHAR. And the project under test has its own autoloader outside the
PHAR. More importantly, atoum can have only one autoloader at a
time. The trick is to generate a temporary file —a temporary autoloader—
that includes the autoloaders from Kitab and from the project under
test. To use the autoloader of Kitab, we have to map/load the PHAR. The
temporary autoloader is then used as the autoloader file for atoum.

And it works. Next commits will optimize this process, but it works.